requirements.txt
pyaml==23.9.6
python-dotenv==1.0.0
PyYAML==6.0.1
scrapli==2023.7.30
